admin cannot delete admin

+++ src/main/java/com/magvy/experis/javalava_backend/domain/services/UserService.java
@@ -82,6 +82,9 @@ public class UserService implements UserDetailsService {
 
     public void deleteUser(Long userId) {
         User user = getUserById(userId);
+        if (user.getRole() == RoleEnum.ADMIN) {
+            throw new IllegalArgumentException("Cannot delete admin user");
+        }
         userRepository.delete(user);
     }
 }
